Working with Razor
In this chapter, we have given you an overview of the Razor view engine
and how it can be used to generate HTML. We showed you how to refer to
data passed from the controller via the view model object and the view
bag, and we showed you how Razor expressions can be used to tailor your
response to the user based on the data you are working with. You will
see many different examples of how Razor can be used in the rest of the
book and we describe how the MVC view mechanism works in detail in
Chapter 18. In the next chapter, we describe the essential development
and testing tools that underpin the MVC Framework and that help you to
get the best from your projects.