Ensential Tools for MVC
In this chapter, we looked at the three tools we find essential for
effective MVC development—Ninject, the built-in Visual Studio support
for unit testing, and Moq. There are many alternatives, both open source
and commercial, for all three tools and you will not lack alternatives
if you do not get along with the tools we like and use. You may find
that you do not like TDD or unit testing in general, or that you are
happy performing DI and mocking manually. That, of course, is entirely
your choice. However, we think there are some substantial benefits in
using all three tools in the development cycle. If you are hesitant to
adopt them because you have never tried them, we encourage you to
suspend disbelief and give them a go—at least for the duration of this
book.