SportStore
Our application, SportsStore, will follow the classic approach taken by
online stores everywhere. We will create an online product catalog that
customers can browse by category and page, a shopping cart where users
can add and remove products, and a checkout where customers can enter
their shipping details. We will also create an administration area that
includes create, read, update, and delete (CRUD) facilities for managing
the catalog—and we will protect it so that only logged-in administrators
can make changes. The application we are going to build is not just a
shallow demonstration. Instead, we are going to create a solid and
realistic application that adheres to best practices. You might find the
going a little slow as we build up the levels of infrastructure we need.
Certainly, you would get the initial functionality built more quickly
with Web Forms, just by dragging and dropping controls bound directly to
a database. But the initial investment in an MVC application pays
dividends, giving us maintainable, extensible, wellstructured code with
excellent support for unit testing. We will be able to speed up things
once we have the basic infrastructure in place.
UNIT